BBBSWA-44 - reorganized some javascript classes into logical folders
